Exception handling - delete

+++ src/main/java/com/dlbrmonteiro/cursospringboot/services/UserService.java
@@ -2,12 +2,15 @@ package com.dlbrmonteiro.cursospringboot.services;
 
 import java.util.List;
 //import java.util.Optional;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.dlbrmonteiro.cursospringboot.entities.User;
 import com.dlbrmonteiro.cursospringboot.repositories.UserRepository;
+import com.dlbrmonteiro.cursospringboot.services.exceptions.DatabaseException;
 import com.dlbrmonteiro.cursospringboot.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -21,11 +24,14 @@ public class UserService {
 	}
 	
 	public User findById(Long id) {
-		//Optional<User> obj = repository.findById(id);
-		// return obj.get();
+//		Optional<User> obj = repository.findById(id);
+//		return obj.get();
 		
-		User user = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
-		return user;
+//		User user = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
+//		return user;
+				
+		Optional<User> obj = repository.findById(id);	
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));				
 	}
 	
 	public User insert(User obj) {
@@ -33,7 +39,16 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+//			if (!repository.existsById(id))
+//				throw new ResourceNotFoundException(id);
+			if (findById(id) != null) 
+				repository.deleteById(id);
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
 	
 	public User update(Long id, User user) {
